Phase 1: fix 64-bit pointer truncation and under-allocation
Pointer-sized values were stored/aligned through 32-bit int in several
places; on LP64 this corrupts pointers and under-allocates pointer arrays.

- p_saveg.c : PADSAVEP + all pointer<->index swizzles use intptr_t.
- r_data.c  : maptexture_t.columndirectory is int (on-disk 4-byte field),
              not void** (was 8 bytes, misaligning patchcount/patches);
              R_InitTextures pointer arrays (textures, texturecolumnlump,
              texturecolumnofs, texturecomposite, ...) allocated with
              sizeof(*array) instead of a hardcoded *4 (half-size on LP64);
              colormap alignment through uintptr_t.
- p_setup.c : sector linebuffer (line_t**) allocated with sizeof(*array)
              instead of *4.
- r_draw.c  : translationtables alignment through uintptr_t.
- z_zone.c  : block->user tag compared as uintptr_t.
- d_net.c   : doomdata_t offset math through intptr_t.
- d_main.c  : -statcopy arg atoi stored via (void*)(intptr_t).

These are genuine 64-bit correctness bugs (the maptexture struct and the
*4 pointer-array allocations were validated by an engine boot smoke test
that crashed before the fixes and boots to the game loop after). In-memory
gameplay/renderer/playsim logic is unchanged.

+++ linuxdoom-1.10/d_net.c
@@ -89,7 +89,7 @@ doomdata_t	reboundstore;
 //
 int NetbufferSize (void)
 {
-    return (int)&(((doomdata_t *)0)->cmds[netbuffer->numtics]); 
+    return (int)((intptr_t)&(((doomdata_t *)0)->cmds[netbuffer->numtics])); 
 }
 
 //
@@ -107,7 +107,7 @@ unsigned NetbufferChecksum (void)
     return 0;			// byte order problems
 #endif
 
-    l = (NetbufferSize () - (int)&(((doomdata_t *)0)->retransmitfrom))/4;
+    l = (NetbufferSize () - (int)((intptr_t)&(((doomdata_t *)0)->retransmitfrom)))/4;
     for (i=0 ; i<l ; i++)
 	c += ((unsigned *)&netbuffer->retransmitfrom)[i] * (i+1);
